requirements.txt
ultralytics
opencv-python-headless
streamlit
numpy
pillow
matplotlib
seaborn
pandas
pyyaml
tqdm
